Added more controllers

+++ controllers/jobs.js
@@ -1,6 +1,4 @@
 const Jobs = require("../model/jobs");
-const jwt = require("jsonwebtoken");
-const validator = require("validator");
 const { StatusCodes } = require("http-status-codes");
 const { badRequest } = require("../errors");
 
@@ -17,39 +15,12 @@ const addJob = async (req, res) => {
 
 const deleteJob = async (req, res) => {
   const { id } = req.params;
+  if (!id) throw new badRequest("Please provide id");
   const job = await Jobs.findOneAndDelete({ _id: id });
   if (!job) throw new badRequest("Job does not Exist");
   res.status(StatusCodes.ACCEPTED).json({ msg: "Job deleted Successfully" });
 };
 
-// const getJobs = async (req, res) => {
-//   const { title, location } = req.query;
-//   queryObject = {};
-//   if (title) {
-//     queryObject.title = { $regex: title, $options: "i" };
-//   }
-
-//   if (location) {
-//     queryObject.location = { $regex: location, $options: "i" };
-//   }
-
-//   let job = Jobs.find(queryObject);
-
-//   if (sort) {
-//     const sortList = sort.split(",").join("");
-//     job = job.sort(sortList);
-//   } else {
-//     result = result.sort("-createdAt");
-//   }
-
-//   const page = Number(req.query.page) || 1;
-//   const limit = Number(req.query.limit) || 15;
-//   const skip = (page - 1) * limit;
-//   job = job.skip(skip).limit(limit);
-
-//   const results = await job;
-//   res.status(StatusCodes.ACCEPTED).json({ results });
-// };
 const getJobs = async (req, res) => {
   const { title, location, sort } = req.query;
   let queryObject = {};
@@ -76,7 +47,7 @@ const getJobs = async (req, res) => {
   const skip = (page - 1) * limit;
   job = job.skip(skip).limit(limit);
 
-  const results = await job.exec();
+  const results = await job;
   res.status(StatusCodes.ACCEPTED).json({ results });
 };
